feat(macos/settings): move TTS/STT cards from Voice to Models & Services
Extract Text-to-Speech and Speech-to-Text configuration into standalone
service cards (TTSServiceCard, STTServiceCard) on the Models & Services
page. Each card uses a Managed/Your Own toggle with Managed disabled
(tooltip: "Managed mode is not provided at this time."). The Voice page
now shows an info banner linking to Models & Services for speech config.
